requirements.txt
# Core optimization and data processing
pulp==2.7.0
pandas>=1.5.0,<2.0.0
numpy>=1.23.0,<2.0.0

# Geospatial and mapping
geopandas>=0.12.0,<1.0.0
global-land-mask>=1.0.0

# API requests
requests>=2.28.0,<3.0.0

# Visualization
plotly>=5.11.0,<6.0.0
kaleido==0.2.1

# Environment variable management
python-dotenv>=1.0.0

# Optional: Jupyter notebook support
jupyter>=1.0.0
notebook>=6.5.0
